chapter09/tiered-model-selection/pyproject.toml
[project]
name = "tiered-model-selection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-core~=0.3",
    "langchain-openai~=0.3",
    "python-dotenv>=1.1.1",
]
